requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
deepface==0.0.93
et_xmlfile==2.0.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
labelImg==1.8.6
libclang==18.1.1
lxml==6.0.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pickle-mixin==1.0.2
pillow==11.3.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==3.17.4
pypdfium2==4.30.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.4
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1

